Overhauled encoder/decoder to be non-linear, optimized image viewer, did readability pass on all files
The pattern generator used by the encoder and decoder is now a generator function and yields x,y pairs.
Besides simplifying the code, this change to the generator allows points in the image to be selected and written/read non-linearly

A quality pass was done on all the files, changing names and sometimes restructuring logic, to make code more readable. Comments are improved.

The image viewer has been significantly optimized. The pygame loop now always waits unless input is given. Now the original image is only ever sampled
to prevent quality loss when scaling the image up from a scaled-down state. The image viewer now automatically launches with an image height of 500px